add support for per instance spki and pfx update callbacks
Since rtrlib itself does not support this, we do this by abusing the
facts that rtrlib does not care about the actual size of a rtr_socket
and that the rtr_socket the update originates from is passed to the
callback function. We can just add some more data after the actual
content of the rtr_socket and access it from the callback function to
find the per instance callback and call it.