Fix error with cffi version 1.16 and newer
cffi seems to have started verifying that the cdef and the c compiler
agree on struct sizes. The sizes for the tr_{tcp,ssh}_config structs did
not match anymore with current rtrlib versions.

Fix this by making the size variable in the cdef so cffi asks the c
compiler about the size.